+++ desapp-groupj-backend/src/main/java/ar/edu/unq/desapp/groupj/backend/model/UserBuilder.java
@@ -8,7 +8,7 @@ public class UserBuilder {
     private String phone = "0000-00000000";
     private String city = "no city";
     private String address = "no address";
-    private Integer credit = 0;
+    private Double credit = 0.0;
 
     public static UserBuilder aUser() {
         return new UserBuilder();
@@ -24,7 +24,7 @@ public class UserBuilder {
         return this;
     }
 
-    public UserBuilder withCredit(int initialCredit) {
+    public UserBuilder withCredit(Double initialCredit) {
         credit = initialCredit;
         return this;
     }
